organise async logic on homepage, style template cards and add images
