+++ packages/server/src/api/controllers/templates.js
@@ -28,7 +28,7 @@ exports.downloadTemplate = async function(ctx) {
 }
 
 exports.exportTemplateFromApp = async function(ctx) {
-  const { appId, instanceId } = ctx.user.appId
+  const { appId, instanceId } = ctx.user
   const { templateName } = ctx.request.body
 
   await exportTemplateFromApp({
